packages/sanmou-common/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sanmou-common"
version = "0.1.0"
description = "Shared game knowledge models and config for Sanmou agents."
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0,<7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sanmou_common = ["config/*.yaml"]
